+++ Docker/build.sh
@@ -1,8 +1,7 @@
 reponame=custardpy
 
-tag=1.5.0
+tag=1.6.0
 docker build -f Dockerfile.$tag -t rnakato/$reponame:$tag . #--no-cache
-
 #docker save -o $reponame-$tag.tar rnakato/$reponame:$tag
 #singularity build -F $reponame.$tag.sif docker-archive://$reponame-$tag.tar
 #exit