+++ tutorial/Hi-C/04.makeEigen.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 
 build=hg38
-fastq_post="_"  # "_" or "_R"  before .fastq.gz
-enzyme=MboI
 gt=genometable.$build.txt
 gene=refFlat.$build.txt
 
-#sing="singularity exec --bind /work,/work2 /work/SingularityImages/custardpy.1.4.3.sif"
+#sing="singularity exec --bind /work,/work2,/work3 /work3/SingularityImages/custardpy.1.5.0.sif"
 sing="singularity exec custardpy.sif"
 
-odir=CustardPyResults_Hi-C/Juicer_$build/Hap1-A
-#odir=CustardPyResults_Hi-C/Juicer_$build/WaplKO_3.3-A
+cell=Control # siCTCF siRad21 siNIPBL
+odir=CustardPyResults_Hi-C/Juicer_$build/$cell
 hic=$odir/aligned/inter_30.hic
+### In case of starting from .hic files:
+#hic=hic/$cell/GSE196034_${cell}_merged.hic
+
 norm=SCALE
 resolution=25000
+ncore=24  # number of cores
 
-$sing makeEigen.sh -p 24 $norm $odir $hic $resolution $gt $gene
+$sing makeEigen.sh -p $ncore $norm $odir $hic $resolution $gt $gene