+++ tutorial/Hi-C/09.4DSimulation.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
-#sing="singularity exec --bind /work,/work2 /work/SingularityImages/custardpy.1.4.3.sif"
+build=hg38
+
+#sing="singularity exec --bind /work,/work2,/work3 /work3/SingularityImages/custardpy.1.6.0.sif"
 sing="singularity exec custardpy.sif"
 
-build=hg38
-odir=CustardPyResults_Hi-C/Juicer_$build/Hap1-A
+cell=Control # siCTCF siRad21 siNIPBL
+odir=CustardPyResults_Hi-C/Juicer_$build/$cell
 hic=$odir/aligned/inter_30.hic
+### In case of starting from .hic files:
+#hic=hic/$cell/GSE196034_${cell}_merged.hic
+
 chr=chr21
 start=24000000
 end=32000000