+++ tutorial/Micro-C/microc.chromap.sh
@@ -7,15 +7,13 @@ gene=refFlat.mm39.txt
 genome=genome.$build.fa
 chromap_index=chromap-indexes/$build
 
-#sing="singularity exec --nv --bind /work,/work2 /work/SingularityImages/custardpy.1.4.3.sif"
-sing="singularity exec custardpy.sif"
+sing="singularity exec --nv --bind /work,/work2,/work3 /work3/SingularityImages/custardpy.1.6.0.sif"
+#sing="singularity exec custardpy.sif"
 
-for cell in ESC_WT01 ESC_WT09
-do
-    $sing custardpy_cooler_MicroC -g $gt -f $genome -i $chromap_index -p $ncore -t chromap fastq/$cell $cell
+cell=C36_rep1
+$sing custardpy_cooler_MicroC -g $gt -f $genome -i $chromap_index -p $ncore -t chromap fastq/$cell $cell
 
-    odir=CustardPyResults_MicroC/$cell/chromap
-    hic=$odir/hic/contact_map.q30.hic
-    norm=SCALE
-    $sing custardpy_process_hic -p $ncore -n $norm -g $gt -a $gene $hic $odir
-done
+odir=CustardPyResults_MicroC/$cell/chromap
+hic=$odir/hic/contact_map.q30.hic
+norm=SCALE
+$sing custardpy_process_hic -p $ncore -n $norm -g $gt -a $gene $hic $odir